DCReportGroup.cs 内のクラスを複数に分割する
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ingen084/KyoshinEewViewerIngen?shareId=XXXX-XXXX-XXXX-XXXX).
